+++ src/browser/index.js
@@ -5,13 +5,13 @@ import Page from '../components/page.js';
 import RawMetadata from '../rawmetadata.js';
 import {GROUPED_METADATA_URL} from '../config.js';
 
-function _renderInBrowser(err, metadataJson, showWorkshopBanner=false) {
+function _renderInBrowser(err, metadataJson) {
   if (err) {
     console.log(err);
   } else {
     const targetNode = document.getElementById('app');
     const allKataGroups = RawMetadata.toKataGroups(metadataJson);
-    React.render(<Page kataGroups={allKataGroups} showWorkshopBanner={showWorkshopBanner}/>, targetNode);
+    React.render(<Page kataGroups={allKataGroups}/>, targetNode);
   }
 }